requirements.txt
absl-py==1.0.0
amqp==5.1.0
arrow==1.2.2
asgiref==3.5.0
astunparse==1.6.3
billiard==3.6.4.0
black==22.3.0
cachetools==5.0.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.3.2
Django==4.0.3
django-icons==21.3
djangorestframework==3.13.1
Flask==2.1.1
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kombu==5.2.4
libclang==13.0.0
line-profiler==3.5.1
make==0.1.6.post2
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.2
MouseInfo==0.1.3
multi-key-dict==2.0.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
numpy==1.22.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-jenkins==1.7.0
pytweening==1.0.4
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.0
wrapt==1.14.0
xgboost==1.5.2
xlrd==2.0.1
